composite_density

+++ tests/user_examples/ExtraSources/extra_sources_share/composite_material.cpp
@@ -9,10 +9,10 @@ namespace SPH
 	void CompositeMaterial::initializeLocalParameters(BaseParticles* base_particles)
 	{
 		ElasticSolid::initializeLocalParameters(base_particles);
-		for (int i = 0; i < CompositeMaterails_.size(); ++i)
+		for (size_t i = 0; i < CompositeMaterails_.size(); ++i)
 			CompositeMaterails_[i]->initializeLocalParameters(base_particles);
 
-		for (int j = 0; j < CompositeMaterails_.size(); ++j)
+		for (size_t j = 0; j < CompositeMaterails_.size(); ++j)
 			sound_speed_.push_back(CompositeMaterails_[j]->ReferenceSoundSpeed());
 
 		c0_ = *std::max_element(sound_speed_.begin(), sound_speed_.end());
